requirements.txt
appnope 
asttokens 
click==8.3.1
comm 
debugpy 
decorator 
executing 
ipykernel 
ipython 
ipython_pygments_lexers 
jedi 
joblib 
jupyter_client 
jupyter_core 
matplotlib-inline 
nest_asyncio 
nltk==3.9.2
numpy 
packaging 
pandas 
parso 
pexpect 
platformdirs 
prompt_toolkit 
psutil 
ptyprocess 
pure_eval 
Pygments 
python-dateutil 
pyzmq 
regex==2026.1.15
scikit-learn 
scipy 
six 
stack_data 
threadpoolctl 
tornado 
tqdm==4.67.3
traitlets 
typing_extensions 
wcwidth 
tensorflow-macos
tensorboard
matplotlib
pandas
streamlit